Changed build script
